% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.GGM_bootstrap.R
\name{coef.estimate}
\alias{coef.estimate}
\alias{coef.GGM_bootstrap}
\title{Precision Matrix to Multiple Regression}
\usage{
\method{coef}{GGM_bootstrap}(object, node = 1, ci = 0.95)
}
\arguments{
\item{object}{object of class \code{estimate} (analytic = F)}

\item{ci}{confidence interval used in the summary output}
}
\value{
list of class \code{coef.estimate}:
}
\description{
There is a direct correspondence between the covariance matrix and multiple regression. In the case of GGMs, it is possible
to estimate the edge set with multiple regression (i.e., neighborhood selection). In *GGMnonreg*, the precision matrix is first sampled from, and then
each draws is converted to the corresponding coefficients and error variances. This results in a posterior distribution. This function can be used
to perform multiple regression.
}
\examples{
# data
X <- scale(GGMnonreg::ptsd)
# fit model
fit <- GGM_bootstrap(X)

# summary for predicting the first variable
coef(fit, node = 1)
}
