% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis.R
\name{confint.ggm_inference}
\alias{confint.ggm_inference}
\title{Extract Confidence Intervals from \code{ggm_inference} Objects}
\usage{
\method{confint}{ggm_inference}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ggm_inference}.}

\item{...}{Currently ignored.}
}
\value{
A matrix including bootstrap confidence intervals.
}
\description{
Extract Confidence Intervals from \code{ggm_inference} Objects
}
\examples{
\donttest{
#  data
Y <- ptsd

# eip
fit <- ggm_inference(Y, method = "spearman",
boot = TRUE, B = 100)

# cis
confint(fit)

}
}
