% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic.R
\name{Diagnostic}
\alias{Diagnostic}
\title{Title}
\usage{
Diagnostic(ThetaEst, Theta)
}
\arguments{
\item{ThetaEst}{Estimated adjacency  matrix}

\item{Theta}{True adjacency  matrix}
}
\value{
Edge identification diagonstic (Specificity, Sensitivity, Fallout, Precision, MCC)
}
\description{
Title
}
