% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{eip}
\alias{eip}
\title{Edge Inclusion "Probability"}
\usage{
eip(Y, method = "pearson", B = 1000, progress = TRUE)
}
\arguments{
\item{Y}{The data matrix of dimensions \emph{n} (observations) by \emph{p} (nodes).}

\item{method}{Character string. Which type of correlation coefficients
to be computed. Options include \code{"pearson"} (default),
\code{"kendall"}, \code{"spearman"}, and \code{"polychoric"}.}

\item{B}{Integer. Number of bootstrap replicates (defaults to \code{1000}).}

\item{progress}{Logical. Should a progress bar be included (defaults to \code{TRUE})?}
}
\value{
An object of class \code{eip}, including a matrix of edge inclusion
"probabilities".
}
\description{
Compute the proportion of bootstrap samples that each relation was selected,
 corresponding to an edge inclusion "probability".
}
\details{
The order is the upper-triangular.
}
\note{
In the context of regression, this general approach was described in
see Figure 6.4. \insertCite{@see Figure 6.4, @Hastie2015;textual}{GGMnonreg}. In this case,
the selection is based on classical hypothesis testing instead of L1-regularization.
}
\examples{
\donttest{
# data
Y <- ptsd

# eip
fit_eip <- eip(Y, method = "spearman")

# print
fit_eip
}
}
\references{
\insertAllCited{}
}
