% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr.R
\name{enr}
\alias{enr}
\title{Expected Network Replicability}
\usage{
enr(net, n, replications, type = "pearson")
}
\arguments{
\item{net}{true network}

\item{n}{samples size}

\item{replications}{number of networks}

\item{type}{pearson or spearman}
}
\value{

}
\description{
Expected Network Replicability
}
\examples{
# correlations
cors <- cor(GGMnonreg::ptsd)

# inverse
inv <- solve(cors)

# partials
pcors <-  -cov2cor(inv)

# set values to zero
pcors <- ifelse(abs(pcors) < 0.05, 0, pcors )

enr(net = pcors, n = 500, replications = 2)
}
