% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMboot.R
\name{GGM_bootstrap.default}
\alias{GGM_bootstrap.default}
\title{Network with a Non-parametric Bootstrap}
\usage{
\method{GGM_bootstrap}{default}(X, iter = 1000, alpha = 0.05)
}
\arguments{
\item{X}{data matrix (dimensions n by p)}

\item{iter}{number of bootstrap samples}

\item{alpha}{desired type I error rate (correspondig to approximately 1 - specificity)}
}
\value{
list of class GGM_bootstrap
\itemize{
\item \code{pcor_selected} selected partial correlation matrix
\item \code{adj_selected} adjacency matrix
\item \code{pcor_mean} saturated matrix (bootstrap mean)
\item \code{boot_results} matrix of bootstrap samples
\item \code{dat} data
}
}
\description{
Network with a Non-parametric Bootstrap
}
\examples{

# data
X <- GGMnonreg::ptsd

# fit model
fit <- GGM_bootstrap(X)

# selected partials
fit$pcor_selected

# selected adjacency
fit$adj_selected

# plot results
plot(fit)
}
