% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMboot.R
\name{GGM_bootstrap}
\alias{GGM_bootstrap}
\title{S3 GGM_bootstrap method}
\usage{
GGM_bootstrap(...)
}
\arguments{
\item{...}{currently not used}
}
\description{
S3 GGM_bootstrap method
}
\seealso{
\code{\link{GGM_bootstrap.default}}
}
