% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare.R
\name{ggm_compare}
\alias{ggm_compare}
\title{Compare Gaussian Graphical Models}
\usage{
ggm_compare(
  Y_g1,
  Y_g2,
  alpha = 0.05,
  control_precision = NULL,
  precision = 0.9
)
}
\arguments{
\item{Y_g1}{data matrix or data frame of dimensions n by p (group 1)}

\item{Y_g2}{data matrix or data frame of dimensions n by p (group 1)}

\item{alpha}{desired type I error rate (corresponding to approximately 1 - specificity)}

\item{control_precision}{should precision be controlled ?
This corresponds to controlling 1 - the false discovery rate (FDR). Default is \code{NULL}.}

\item{precision}{desired precision (1 - FDR). Default is 0.90 (1 - 0.9 = FDR of 0.1).}
}
\value{
list containing the adjacency matrix (0's and 1's), partial correlation differences, and an
'adjacency' matrix containing the partial correlation differences.
}
\description{
Compare Gaussian Graphical Models
}
\examples{
true <- gen_pcors(20, edge_prob = 0.3)
Y1 <- MASS::mvrnorm(100,
                    mu = rep(0, 20),
                    Sigma = true$cors)

Y2 <- MASS::mvrnorm(100,
                   mu = rep(0, 20),
                   Sigma = diag(20))

fit <- ggm_compare(Y1, Y2, precision = 0.80,
                   control_precision = TRUE)
}
