% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcor_fisher_z.R
\name{GGM_fisher_z}
\alias{GGM_fisher_z}
\title{S3 GGM_fisher_z method}
\usage{
GGM_fisher_z(...)
}
\arguments{
\item{...}{currently not used}
}
\description{
S3 GGM_fisher_z method
}
\seealso{
\code{\link{GGM_fisher_z.default}}
}
