% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMregress.R
\name{GGM_regression.default}
\alias{GGM_regression.default}
\title{Network with Multiple Regression (Ordinary Least Squares)}
\usage{
\method{GGM_regression}{default}(X, IC = "BIC", method = "exhaustive",
  rule = "and", ...)
}
\arguments{
\item{X}{data Matrix (dimensions n by p)}

\item{IC}{information criterion (\code{AIC} or \code{BIC})}

\item{method}{subsect selection method (\code{forward}, \code{backward}, or \code{exhaustive})}

\item{rule}{decision rule (see notes)}

\item{...}{currently ignored}
}
\description{
Network with Multiple Regression (Ordinary Least Squares)
}
\note{
The \code{rule} arguement can either be \code{or} or \code{and}. For the former, only the edge only needs to
be selected in one regrssion models, whereas, for the later, each edge must be selected in both regression models.
Setting \code{rule = "and"} should result in a network with fewer connections.
}
\examples{
# data
X <- GGMnonreg::ptsd

# fit model
fit <- GGM_regression(X)

# selected partials
fit$pcor_selected

# selected adjacency
fit$adj_selected
}
