% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMregress.R
\name{GGM_regression}
\alias{GGM_regression}
\title{S3 GGM_regression method}
\usage{
GGM_regression(...)
}
\arguments{
\item{...}{currently not used}
}
\description{
S3 GGM_regression method
}
\seealso{
\code{\link{GGM_regression.default}}
}
