% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_search.R
\name{ggm_search}
\alias{ggm_search}
\title{Gaussian graphical model: automated search}
\usage{
ggm_search(Y, IC = "BIC", method = "forward")
}
\arguments{
\item{Y}{The data matrix of dimensions \emph{n} (observations) by \emph{p} (nodes).}

\item{IC}{Character string. The desired information criterion. Options include
\code{"AIC"} and \code{"BIC"} (default).}

\item{method}{Character string. The desired subsect selection method
Options includes \code{"forward"} (default), \code{"backward"},
and \code{"exhaustive"}.}
}
\value{
An object of class \code{ggm_search}.
}
\description{
Gaussian graphical model: automated search
}
\note{
Multiple regression is used to estimate the graph.
}
\examples{
\donttest{
# data
Y <- ptsd

# search data
fit <- ggm_search(Y)
}

}
