% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMregress.R
\name{GGMregress}
\alias{GGMregress}
\title{Title}
\usage{
GGMregress(X, IC, method)
}
\arguments{
\item{X}{Data Matrix (dimensions n by p)}

\item{IC}{Information criteria (AIC or BIC)}

\item{method}{Subsect selection method}
}
\description{
Title
}
