% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ising_search.R
\name{ising_search}
\alias{ising_search}
\title{Ising: automated search}
\usage{
ising_search(Y, IC = "BIC", progress = TRUE)
}
\arguments{
\item{Y}{A data matrix of dimensions \emph{n} (observations) by \emph{p} (nodes).}

\item{IC}{Character string. The desired information criterion. Options include
\code{"AIC"} and \code{"BIC"} (default).}

\item{progress}{Logical. Should a progress bar be included (defaults to \code{TRUE})?}
}
\value{
An object of class \code{ising_search} including:

\itemize{

\item{\strong{wadj}}: Weighted adjacency matrix, corresponding
to the partial correlation network.

\item{\strong{adj}}: Adjacency matrix (detected effects).

\item{\strong{pcors}}: Partial correlations.

\item{\strong{n}}: Sample size.

\item{\strong{p}}: Number of nodes.

\item{\strong{Y}}: Data.

}
}
\description{
Data mining to learn the graph of binary variables with an Ising model
             \insertCite{lenz1920beitrvsge,ising1925beitrag}{GGMnonreg}.
}
\details{
Currently only backwards selection is currently implemented.
}
\note{
For an excellent overview of the Ising model see \insertCite{marsman2018introduction;textual}{GGMnonreg}.
}
\examples{
\donttest{
# data
Y <- ifelse( ptsd[,1:5] == 0, 0, 1)

# search data
fit <- ising_search(Y)
}
}
\references{
\insertAllCited{}
}
