% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Maximum Likelihood (Inverse) Covariance Matrix}
\usage{
mle(X)
}
\arguments{
\item{X}{Data Matrix (dimensions n by p)}
}
\value{
inv_cov_mle MLE precision matrix

cov_mle MLE covariance matrix
}
\description{
Maximum Likelihood (Inverse) Covariance Matrix
}
\examples{
# data
X <- GGMnonreg::ptsd

mle(X)
}
