% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor_summary.R
\name{pcor_summary.GGM_bootstrap}
\alias{pcor_summary.GGM_bootstrap}
\title{Summary Method for a \code{pcor_summary.GGM_bootstrap} object}
\usage{
\method{pcor_summary}{GGM_bootstrap}(object, ...)
}
\arguments{
\item{object}{object of class \code{pcor_summary.GGM_bootstrap}}

\item{...}{currently ignored}
}
\value{
data frame of partial correlations with confidence intervals
}
\description{
Summary Method for a \code{pcor_summary.GGM_bootstrap} object
}
\examples{
# data
X <- GGMnonreg::ptsd

# fit model
fit <- GGM_bootstrap(X)

# summarise the partial correlations
pcor_summ <- pcor_summary(fit)
pcor_summ
}
