% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor_summary.R
\name{pcor_summary.GGM_fisher_z}
\alias{pcor_summary.GGM_fisher_z}
\title{Summary Method for \code{pcor_summary.GGM_fisher_z} objects}
\usage{
\method{pcor_summary}{GGM_fisher_z}(object, ...)
}
\arguments{
\item{object}{object of class \code{pcor_summary.GGM_fisher_z}}

\item{...}{currently ignored}
}
\value{
data frame of partial correlations with confidence intervals
}
\description{
Summary Method for \code{pcor_summary.GGM_fisher_z} objects
}
\examples{
# data
X <- GGMnonreg::ptsd

# fit model
fit <- GGM_fisher_z(X)

# summarise the partial correlations
pcor_summ <- pcor_summary(fit)
pcor_summ
}
