% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor_summary.R
\name{pcor_summary}
\alias{pcor_summary}
\title{S3 pcor_summary method}
\usage{
pcor_summary(object, ...)
}
\arguments{
\item{object}{object of class \code{GGM_fisher_z} or \code{GGM_bootstrap}}

\item{...}{not currently used}
}
\value{
\code{select} works with the following methods:
\itemize{
\item \code{\link{GGM_fisher_z}}
\item \code{\link{GGM_bootstrap}}
}
}
\description{
S3 pcor_summary method
}
