% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr.R
\name{plot.enr}
\alias{plot.enr}
\title{Plot \code{enr} Objects}
\usage{
\method{plot}{enr}(x, iter = 1e+05, fill = "#009E73", alpha = 0.5, ...)
}
\arguments{
\item{x}{Object of class \code{enr}}

\item{iter}{number of samples}

\item{fill}{fill color for density}

\item{alpha}{transparency}

\item{...}{currently ignored}
}
\value{

}
\description{
Plot \code{enr} Objects
}
\examples{
# correlations
cors <- cor(GGMnonreg::ptsd)

# inverse
inv <- solve(cors)

# partials
pcors <-  -cov2cor(inv)

# set values to zero
pcors <- ifelse(abs(pcors) < 0.05, 0, pcors )

est <- enr(net = pcors, n = 500, replications = 2)

# plot
plot(est)
}
