% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMregress.R
\name{plot.GGM_regression}
\alias{plot.GGM_regression}
\title{Plot \code{GGM_regression} Network}
\usage{
\method{plot}{GGM_regression}(x, layout = "circle",
  edge_colors = "classic", node_labels = NULL,
  node_labels_color = "black", node_groups = NULL,
  node_outer_size = 12, node_inner_size = 11, ...)
}
\arguments{
\item{x}{object of class \code{GGM_regression}}

\item{layout}{network layout (\link[sna]{gplot.layout})}

\item{edge_colors}{color theme for positive and negative edges ("classic", "color_blind", or "vivid")}

\item{node_labels}{node labels}

\item{node_labels_color}{node labels color}

\item{node_groups}{node group indicator}

\item{node_outer_size}{node border size}

\item{node_inner_size}{node size}

\item{...}{additional arguments (\link[GGally]{ggnet2})}
}
\value{
object of class \code{ggplot}
}
\description{
Plot \code{GGM_regression} Network
}
\note{
See palette options here \url{http://www.sthda.com/english/wiki/ggplot2-colors-how-to-change-colors-automatically-and-manually#use-rcolorbrewer-palettes}.
}
\examples{
# data
X <- GGMnonreg::ptsd

# fit model
fit <- GGM_regression(X)

# selected partials
fit$pcor_selected

# selected adjacency
fit$adj_selected
plt <- plot(fit,
            node_labels = letters[1:20],
            node_labels_color = "white",
            node_groups = rep(c("1", "2", "3", "4"), each = 5),
            edge_colors = "classic",
            edge.alpha = 0.5, palette = "Set2")
}
