% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor_summary.R
\name{plot.pcor_summary.GGM_fisher_z}
\alias{plot.pcor_summary.GGM_fisher_z}
\title{Plot \code{pcor_summary.GGM_fisher_z} Edges}
\usage{
\method{plot}{pcor_summary.GGM_fisher_z}(object, size = 1,
  color = "black", customize = "GGMnonreg", ...)
}
\arguments{
\item{object}{object of class \object{pcor_summary.GGM_fisher_z }}

\item{size}{point size}

\item{color}{point color}

\item{customize}{\code{"GGMnonreg"} for customized plot and \code{user} for a bare plot to be further
customized}

\item{...}{currently ignored}
}
\value{
a \code{ggplot} object
}
\description{
Plot \code{pcor_summary.GGM_fisher_z} Edges
}
\examples{
# data
X <- GGMnonreg::ptsd

# fit model
fit <- GGM_fisher_z(X)

# summarise the partial correlations
pcor_summ <- pcor_summary(fit)

# plot
plot(pcor_summ)
}
