% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GGM_regression.R
\name{plot.predict.GGM_regression}
\alias{plot.predict.GGM_regression}
\title{Plot \code{predict.GGM_regression} Object}
\usage{
\method{plot}{predict.GGM_regression}(x, ci = 0.95, color = "black",
  size = 1, ...)
}
\arguments{
\item{x}{object of class \code{predict.GGM_regression}}

\item{ci}{confidence interval}

\item{color}{point color}

\item{size}{point size}

\item{...}{currently ignored}
}
\value{
\code{ggplot} object
}
\description{
Plot \code{predict.GGM_regression} Object
}
\examples{
# data
X <- GGMnonreg::ptsd[,1:5]

# fit model
fit <- GGM_regression(X)

# predict
net_pred <- predict(fit, iter = 10)

#plot
plot(net_pred)
}
