% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GGM_regression.R
\name{predict.GGM_regression}
\alias{predict.GGM_regression}
\title{Network Predictability}
\usage{
\method{predict}{GGM_regression}(object, new_data = NULL,
  resample = TRUE, iter = 500, ...)
}
\arguments{
\item{object}{an object of class \code{GGM_regression}}

\item{new_data}{optional new data for out-of-sample predictability}

\item{resample}{bootstrap predictability (see notes)}

\item{iter}{number of bootstrap iteractions}

\item{...}{currently ignored}
}
\value{
list of class GGM_bootstrap
\itemize{
\item \code{R2_boot_storage} matrix of bootstrapped estimates
\item \code{mse_boot_storage} matrix of bootstrapped estimates
\item \code{iter} number of bootstrap samples
\item \code{p} number of nodes
\item \code{n} number of observations
\item \code{newdata} \code{TRUE} or \code{FALSE}
\item \code{reg_object} object of class \code{GGM_regression}
\item \code{resample} \code{TRUE} or \code{FALSE}
}
}
\description{
Network Predictability
}
\note{
Currently it is only possible to bootstrap predictability, that is variance explained and mean squared
error. This provides bootstrapped distribution for each. Computing a point estimate will be implemented
in the future
}
\examples{
# data
X <- GGMnonreg::ptsd[,1:5]

# fit model
fit <- GGM_regression(X)

# predict
net_pred <- predict(fit, iter = 10)
net_pred
}
