% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.GGM_bootstrap.R
\name{print.coef.GGM_bootstrap}
\alias{print.coef.GGM_bootstrap}
\title{Print method for \code{coef.GGM_bootstrap} objects}
\usage{
\method{print}{coef.GGM_bootstrap}(x, ...)
}
\arguments{
\item{x}{An object of class \code{coef.GGM_bootstrap}}

\item{...}{currently ignored}
}
\description{
Print method for \code{coef.GGM_bootstrap} objects
}
\seealso{
\code{\link{coef.GGM_bootstrap}}
}
