% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr.R
\name{print.enr}
\alias{print.enr}
\title{Print Method for \code{enr} Objects}
\usage{
\method{print}{enr}(x, ...)
}
\arguments{
\item{x}{object of class \code{enr}}

\item{...}{currently ignored}
}
\value{

}
\description{
Print Method for \code{enr} Objects
}
