% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcor_fisher_z.R
\name{print.GGM_fisher_z}
\alias{print.GGM_fisher_z}
\title{Print method for a \code{GGM_fisher_z} object}
\usage{
\method{print}{GGM_fisher_z}(x, ...)
}
\arguments{
\item{x}{An object of class \code{GGM_fisher_z}}

\item{...}{currently ignored}
}
\description{
Print method for a \code{GGM_fisher_z} object
}
\examples{
X <- GGMnonreg::ptsd[, 1:5]
fit <- GGM_fisher_z(X)
fit
}
