% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMregress.R
\name{print.GGM_regression}
\alias{print.GGM_regression}
\title{Print method for a \code{GGM_regression} object}
\usage{
\method{print}{GGM_regression}(x, ...)
}
\arguments{
\item{x}{An object of class \code{GGM_regression}}

\item{...}{currently ignored}
}
\description{
Print method for a \code{GGM_regression} object
}
\examples{
X <- GGMnonreg::ptsd[, 1:5]
fit <- GGM_regression(X)
fit
}
