% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmnonreg_print.R
\name{print.ggmnonreg}
\alias{print.ggmnonreg}
\title{Print \code{ggmnonreg} Object}
\usage{
\method{print}{ggmnonreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggmnonreg}}

\item{...}{Currently ignored}
}
\value{
No return value.
}
\description{
Print \code{ggmnonreg} Object
}
