% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor_summary.R
\name{print.pcor_summary.GGM_fisher_z}
\alias{print.pcor_summary.GGM_fisher_z}
\title{Print Method for \code{pcor_summary.GGM_fisher_z} objects}
\usage{
\method{print}{pcor_summary.GGM_fisher_z}(x, ...)
}
\arguments{
\item{x}{object of class \code{pcor_summary.GGM_fisher_z}}

\item{...}{currently ignored}
}
\description{
Print Method for \code{pcor_summary.GGM_fisher_z} objects
}
\examples{
# data
X <- GGMnonreg::ptsd

# fit model
fit <- GGM_fisher_z(X)

# summarise the partial correlations
pcor_summ <- pcor_summary(fit)
pcor_summ
}
