% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GGM_regression.R
\name{print.predict.GGM_regression}
\alias{print.predict.GGM_regression}
\title{Print method for a \code{predict.GGM_regression} object}
\usage{
\method{print}{predict.GGM_regression}(x, ...)
}
\arguments{
\item{x}{object of class \code{predict.GGM_regression}}

\item{...}{currently ignored}
}
\description{
Print method for a \code{predict.GGM_regression} object
}
\examples{
# data
X <- GGMnonreg::ptsd[, 1:5]

# fit model
fit <- GGM_regression(X)

# predict
net_pred <- predict(fit, iter = 10)
net_pred
}
