% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GGM_regression.R
\name{print.summary.predict.GGM_regression}
\alias{print.summary.predict.GGM_regression}
\title{print method for object of class \code{summary.predict.GGM_regression}}
\usage{
\method{print}{summary.predict.GGM_regression}(x, ...)
}
\arguments{
\item{x}{object of class \code{summary.predict.GGM_regression}}

\item{...}{currently ignored}
}
\description{
print method for object of class \code{summary.predict.GGM_regression}
}
\examples{
# data
X <- GGMnonreg::ptsd[, 1:5]

# fit model
fit <- GGM_regression(X)

# predict
net_pred <- predict(fit, iter = 10)
net_pred_summ <- summary(net_pred)
net_pred_summ
}
