% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_2_inverse.R
\name{regress_2_inverse}
\alias{regress_2_inverse}
\title{regression to inverse covariance matrix}
\usage{
regress_2_inverse(x)
}
\arguments{
\item{x}{data matrix (dimensions n times p)}
}
\value{
inverse-cov Inverse of the covariance matrix (i.e., the precision matrix)

cov_matrix Covariance matrix
}
\description{
regression to inverse covariance matrix
}
\examples{
p = 10
n = 100
x <- matrix(rnorm(p * n), nrow = n, ncol = p)

estimate <- regress_2_inverse(x)

all.equal(solve(cov(x)), estimate$inverse_cov)


}
