% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr.R
\name{summary.enr}
\alias{summary.enr}
\title{Summary Method for \code{enr} Objects}
\usage{
\method{summary}{enr}(object, ...)
}
\arguments{
\item{object}{object of class \code{enr}}

\item{...}{currently ignored}
}
\value{

}
\description{
Summary Method for \code{enr} Objects
}
