% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcor_fisher_z.R
\name{summary.GGM_fisher_z}
\alias{summary.GGM_fisher_z}
\title{Summary method for a \code{GGM_fisher_z} object}
\usage{
\method{summary}{GGM_fisher_z}(object, ...)
}
\arguments{
\item{object}{An object of class \code{GGM_fisher_z}}
}
\description{
Summary method for a \code{GGM_fisher_z} object
}
\examples{
X <- GGMnonreg::ptsd[, 1:5]
fit <- GGM_fisher_z(X)
summary(fit)
}
