% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMregress.R
\name{summary.GGM_regression}
\alias{summary.GGM_regression}
\title{Summary method for a \code{GGM_regression} object}
\usage{
\method{summary}{GGM_regression}(object, ...)
}
\arguments{
\item{object}{An object of class \code{GGM_regression}}

\item{...}{currently ignored}
}
\description{
Summary method for a \code{GGM_regression} object
}
\examples{
X <- GGMnonreg::ptsd[, 1:5]
fit <- GGM_regression(X)
summary(fit)
}
