% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GGM_regression.R
\name{summary.predict.GGM_regression}
\alias{summary.predict.GGM_regression}
\title{Summary method for a \code{GGM_regression} object}
\usage{
\method{summary}{predict.GGM_regression}(object, ci = 0.95, ...)
}
\arguments{
\item{object}{object of class \code{GGM_regression}}
}
\value{
data frame(s) including the summarized bootstrap samples
}
\description{
Summary method for a \code{GGM_regression} object
}
\examples{
# data
X <- GGMnonreg::ptsd[, 1:5]

# fit model
fit <- GGM_regression(X)

# predict
net_pred <- predict(fit, iter = 25)
net_pred_summ <- summary(net_pred)
net_pred_summ
}
