% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained.R
\name{constrained}
\alias{constrained}
\title{Constrained Precision Matrix}
\usage{
constrained(Sigma, adj)
}
\arguments{
\item{Sigma}{Covariance matrix}

\item{adj}{Matrix with constraints. A zero indicates that element
should be constrained to zero.}
}
\value{
A list containing the inverse covariance matrix and the covariance matrix.
}
\description{
Compute the maximum likelihood estimate of the precision matrix,
given a known graphical structure (i.e., an adjacency matrix).
This approach was originally described in
"The Elements of Statistical Learning"
\insertCite{@see pg. 631, @hastie2009elements}{GGMnonreg}.
}
\note{
The algorithm is written in \code{c++}, and should scale to high dimensions.

Note there are a variety of algorithms for this purpose. Simulation
studies indicated that this approach is both accurate and computationally
efficient \insertCite{@HFT therein, @emmert2019constrained}{GGMnonreg}
}
\examples{
# data
Y <- ptsd

# estimate graph
fit <- ggm_inference(Y, boot = FALSE)

# constrain to zero
constrained_graph <- constrained(cor(Y), fit$adj)

}
\references{
\insertAllCited{}
}
