% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{get_graph}
\alias{get_graph}
\title{Get Graph}
\usage{
get_graph(x)
}
\arguments{
\item{x}{An object of class \code{ggmnonreg}}
}
\value{
A list including two matrices (the weighted adjacency and adjacency matrices)
}
\description{
Extract the necessary ingredients to visualize the conditional
             dependence structure.
}
\examples{
# data
Y <- ptsd

# estimate graph
fit <- ggm_inference(Y, boot = FALSE)

# get info for plotting
get_graph(fit)
}
