% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMboot.R
\name{GGMboot}
\alias{GGMboot}
\title{Non-parametric bootstrap}
\usage{
GGMboot(X, sims, alpha)
}
\arguments{
\item{X}{Data matrix (dimensions n \times p)}

\item{sims}{number of bootstrap samples}

\item{alpha}{type I error rate (correspondig to approximately 1 - specificity)}
}
\value{
mat_selected adjacency matrix

mat_mean mean of the bootstrap samples

boot_results bootstrap samples
}
\description{
Non-parametric bootstrap
}
\examples{

fit <- GGMboot(X, sims = 1000, alpha = 0.01)

# adjacency matrix
fit_boot$mat_selected
}
