% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMboot.R
\name{GGMboot}
\alias{GGMboot}
\title{Title}
\usage{
GGMboot(X, sims, alpha)
}
\arguments{
\item{X}{Data matrix (n times p)}

\item{sims}{number of bootstrap samples}

\item{alpha}{type I error rate (correspondig to 1 - specificity)}
}
\description{
Title
}
