% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ising_search.R
\name{ising_search}
\alias{ising_search}
\title{Ising: automated search}
\usage{
ising_search(Y, IC = "BIC", progress = TRUE)
}
\arguments{
\item{Y}{A data matrix of dimensions \emph{n} (observations) by \emph{p} (nodes).}

\item{IC}{Character string. The desired information criterion. Options include
\code{"AIC"} and \code{"BIC"} (default).}

\item{progress}{Logical. Should a progress bar be included (defaults to \code{TRUE})?}
}
\value{
An object of class \code{ising_search}, including \code{wadj} (weighted adjacency matrix)
        and \code{adj} (adjacency matrix).
}
\description{
Data mining to learn the graph.
}
\details{
Only backwards selection is currently implemented.
}
\examples{
\donttest{
# data
Y <- ifelse( ptsd[,1:5] == 0, 0, 1)

# search data
fit <- ising_search(Y)
}
}
