% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Maximum likelihood precision and covariance matrices}
\usage{
mle(X)
}
\arguments{
\item{X}{Data Matrix (dimensions n by p)}
}
\value{
inv_cov_mle MLE precision matrix

cov_mle MLE covariance matrix
}
\description{
Maximum likelihood precision and covariance matrices
}
